Switched the analog sensors back to using uint16_t for the sample datatype. The reason this was switched was to accomodate ratiometric sensors, however introducing the sampleVdd parameter solved the issue without requiring the datatype change.

+++ src/peripherals/analog_linear.h
@@ -35,7 +35,7 @@ typedef struct
 {
 	ANALOG_SENSOR_FIELDS;
 	const linearSensorConfig_t*	config;
-	float sample;
+	uint16_t sample;
 	float value;
 } linearSensor_t;
